Update front and backend logics
