+++ src/main/resources/templates/reservation.html
@@ -325,7 +325,7 @@
 				
 				//ボックスタグ作成
 				const boxTag = `<div class="w-full sm:w-1/3">
-					<input type="checkbox" id="${softcode}" name="soft" value="${softcode}" />
+					<input type="radio" id="${softcode}" name="soft" value="${softcode}" />
     				<label for="${softcode}" class="text-sm sm:text-base">${softname}</label>
 				</div>`;
 				
@@ -342,9 +342,11 @@
 	
 	//機種コードの取得[チェックボックス]
 	function clickCheckbox() {
+		// 階数の取得
+		const floor = document.floorform.floor.value;
 		
 		//チェックボックスの値取得
-	    var softcode = '';
+	    let softcode = '';
 	    const checkbox_soft = document.soft_form.soft;
 
 	    for (let i = 0; i < checkbox_soft.length; i++) {
@@ -357,30 +359,34 @@
 	    $.ajax({
 			type : "GET",
 			url : "/get_machine/fromsoft",
-			data : {js_softcode : softcode},
+			data : {js_softcode : softcode, js_floor: floor},
 			dataType : "json"
 		})
 		.then(function(response_data) {
-			// 機種コードのコンテナ
+			// UIコンポーネントの取得
 			const btnContainer = $("#btnContainer");
 			const filterBtn = $("#filterBtn");
+			const backdrop = $("#backdrop");
 			
 			// コンテナの初期化
 			btnContainer.empty();
 			
 			// 動的表示
-			for(i = 0; i < response_data.length; i++) {
-				const machinecode = (response_data[i]);
-				const btnTag = `<button name='${machinecode}' th:text='${machinecode}' th:value='${machinecode}' th:onclick='clickMachine()' class='w-full px-1 h-10 sm:h-12 border hover:bg-green-200 hover:text-white focus:ring ring-gray-100 rounded-lg text-sm text-black'>${machinecode}</button>`;
-								
-				//表示する場所の取得
-				btnContainer.append(btnTag);
-			}
+			if (response_data.length > 0) {
+				for(i = 0; i < response_data.length; i++) {
+					const machinecode = (response_data[i]);
+					const btnTag = `<button name='${machinecode}' th:text='${machinecode}' th:value='${machinecode}' th:onclick='clickMachine()' class='w-full px-1 h-10 sm:h-12 border hover:bg-green-200 hover:text-white focus:ring ring-gray-100 rounded-lg text-sm text-black'>${machinecode}</button>`;
+									
+					//表示する場所の取得
+					btnContainer.append(btnTag);
+				}
+			} else {
+				const blankTag = `<span class="text-sm">該当無し</span>`;
+				btnContainer.append(blankTag);
+			}	
 			
-			// 絞り込みボタンの活性化
-			filterBtn.prop("disabled", false);
-			filterBtn.removeClass("bg-gray-100 text-gray-200 cursor-not-allowed");
-			filterBtn.addClass("bg-gray-200 hover:bg-gray-400 hover:text-white focus:ring ring-gray-100 text-black");
+			// モーダルの非表示
+			backdrop.addClass("hidden");
 		}
 		, function() {
 			//失敗時
